Move CCryptoKeyStore to crypter.cpp
This breaks the dependency on crypter for disable-wallet builds.
